+++ kpi_extraction/stabililty_metrics.py
@@ -16,7 +16,7 @@ def get_ct_stability(cycle_times):
     # normalized stability score
     ct_stability_index = 1 / (1 + ct_stability)
 
-    return ct_stability_index
+    return float(ct_stability_index)
 
 """
 Returns a activity time stability index (mean)
@@ -31,7 +31,7 @@ def get_activity_time_stability(data):
 
     activity_durations_vc = 1 / (1 + activity_durations_vc)
 
-    return activity_durations_vc.mean()
+    return float(activity_durations_vc.mean())
 
 """
 Returns the amount of distinc variants and ap path stability index (shannon entropy)
@@ -52,5 +52,5 @@ def get_path_stability(data):
 
     return {
         "variant-amount": len(variants_count),
-        "index": path_stability_index
+        "index": float(path_stability_index)
     }
